remove catalogs
- Glue
- Rest

This removes imports of opencensus that leaks a go routine
